requirements.txt
google-genai
filetype
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
pydantic-settings
sqlalchemy
psycopg2-binary